requirements.txt
discord.py==2.1.0
PyNaCl==1.5.0
yt-dlp==2023.01.02
Flask==2.2.3
